Log Size and Performance Improvements (#212)
About 10GB was freed on my laptop from just Protobuf logs. This was
caused by constantly logging the exact same data over and over again.

Instead, it will only log non-duplicate data, if the same message is
logged twice in a row, it will ignore the newest one.

This reduced the size of logs while connected to the rover by around 80%

Additionally, when flushing the log buffer, all files are written in
parallel to improve performance

---------

Co-authored-by: lregueiferos <124931111+lregueiferos@users.noreply.github.com>